Moved the endpoints to the controllers directory. This is to do a MVC style for the backend. Also added a couple extra endpoints. Still need to add the logic. Might look into firebase.
